remove duplicates of the protocol ids
