+++ apps/data_handler/db/models/liquidable_debt.py
@@ -5,7 +5,7 @@ from sqlalchemy import DECIMAL, BigInteger, Column, String
 from sqlalchemy_utils.types.choice import ChoiceType
 
 from shared.db import Base
-from shared.constants import ProtocolIDs
+from shared.protocol_ids import ProtocolIDs
 
 
 class LiquidableDebt(Base):
