+++ apps/data_handler/handlers/health_ratio_level/nostra_alpha.py
@@ -8,7 +8,7 @@ from data_handler.handlers.liquidable_debt.values import (
 from health_ratio_handlers import NostrAlphaHealthRatioHandler
 
 from data_handler.db.models import HealthRatioLevel
-from shared.constants import ProtocolIDs
+from shared.protocol_ids import ProtocolIDs
 
 
 def run():
